fix(locationLists): fix methods for retrieving/updating unresolved locations
Swagger docs for the following two endpoints were incorrect:
  1) PUT /v1/location_lists/{location_list_id}/upload/{file_upload_id}/unresolved_locations
  2) PUT /v1/location_lists/{location_list_id}/upload/{file_upload_id}/unresolved_locations/{location_id}

1) Although this endpoint does support PUT, there was no documentation describing
the body of the request. After reviewing the Gigwalk API, I believe this should
be documenting the GET operation. This makes sense as there is currently no swagger
doc for that method.

2) Technically, there is nothing wrong with the documentation for this method, but
it's a bit misleading. The location_id parameter is not used anywhere in the API
source code, so it should be removed altogether. I updated our client to not even
expose this parameter.