fix(requestQueue): copy all config options to request object and stringify the request
Due to a flow issue, we can't use object destructuring to copy axios config
options to the request object AND have it type check. A workaround is to manually
check for the existence of an option and copy it to the request object if it
exists. I made this change earlier but forgot to check for baseConfig options
(ex. headers, responseType, etc)

Also, we should stringify the same object that checkQueue will stringify to
ensure the generated key is correct.