fix(api-tests): omit transform-runtime plugin when running mocha tests
There seems to be a conflict between the transform-runtime and rewire
plugins for babel. When both are included, running `npm test` throws
a rather cryptic error. The stacktrace points to rewire being the
culprit, but tests run just fine when you remove transform-runtime.

Since we really only need the transform-runtime plugin to be enabled
for production builds, I updated .babelrc to include it when NODE_ENV
is set to production